+++ src/content/content.test.ts
@@ -22,6 +22,7 @@ describe('content module', () => {
       disconnect() {}
       constructor() {}
     } as any;
+    document.documentElement.setAttribute('lang', 'ja');
   });
 
   afterEach(() => {
@@ -32,6 +33,7 @@ describe('content module', () => {
     }
     global.MutationObserver = originalMutationObserver as any;
     document.body.innerHTML = '';
+    document.documentElement.removeAttribute('lang');
   });
 
   test('clickSkipButton clicks the skip intro button if present', async () => {
